+++ src/tools/map_compiler/export_obj.h
@@ -29,7 +29,7 @@ bool ExportObj(const Q3::Map& map, const char* path) {
 
 		for (size_t i=0; i<brush.planes.size(); ++i) {
 			const size_t MAX_FACE_EDGES = 256;
-			u8 face_edges[MAX_FACE_EDGES];
+			u32 face_edges[MAX_FACE_EDGES];
 			size_t num_face_edges = 0;
 
 			u32 face_mask = 1 << i;
